requirements.txt
tensorflow==1.12.0
Flask==1.1.1
plotly==4.6.0
dash_html_components==1.0.3
Keras==2.2.4
numpy==1.16.1
Werkzeug==0.16.0
dash==1.10.0
dash_core_components==1.9.0
pandas==1.0.3
gevent==1.4.0
python_dateutil==2.8.1
